Client: Add snap info to each task
In order to allow Refresh Awareness to be able to link a task
to the corresponding snap being refreshed when a single notice
contains several snaps, it is a must to include that info into
each task.

This patch adds the snap name, snap instance and snap revision
(when available) into each task.