i/prompting: remove NewIDAndTimestamp
IDs used for prompting (rules and prompt requests) must be unique across
reboots, and thus cannot rely on timestamps or even monotonic time. We
also want IDs to be sorted, so we cannot use boot ID along with
monotonic time either. Lastly, since we need IDs to be unique across
reboots, we can't simply use a counter based on the maximum of the
existing rule/prompt IDs. Instead, we should store the current maximum
ID in memory and on disk alongside the storage of rules and prompts.
This should be implemented in the respective forthcoming packages and
used at the former callsights of NewIDAndTimestamp.

Since IDs should no longer be tied to timestamps, we can simply use
`time.Now()` where the timestamps from `NewIDAndTimestamp` were formerly
used.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>