cmd/snap-confine: remove unused spread tests
When the snap-confine repository was merged into snapd, circa 2016, it brought
its own collection of spread tests but due to the differences in the
project-wide setup code, they could not just be enabled and used directly.

Over time some tests were "upgraded" and moved to the main repository but the
bulk has stayed behind, in unchanged state and no longer matching the new
behavior of snap-confine.

Since we are trying to improve the quality of existing tests, let's remove
those unused tests so that we don't spend time on them by accident.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>