tests/completion: document the "three" completion tests
There are "three" completion tests that multiply with the large number of
variatns present in the suite. The "simple" test is unsual and I don't fully
understand how it works because it does not install the test-snapd-complexion
snap at all. Instead it relies on running the completion script directly,
without the snap complete/etelpmoc script pair working across the sandbox.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>