o/snapstate: apps don't wait for reboot on classic w/ kernel (#13852)
Refreshes of essential snaps are organised to have a single reboot.
However, on classic systems with kernel/gadget snaps this means that
apps must wait for the essential snaps and therefore for a reboot. To
prevent apps from blocking until then, this change splits a refresh into
two tasksets. One with essential snaps which is still organised
according to the single reboot and another with apps and their bases
which can complete before the reboot. Some constraints on this are snapd
which is always refreshed first and apps that depend on the model base.
In those cases, we add dependencies across essential/non-essential sets.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>