notify: Don't show complete refresh notification if refresh-observe is connected
The new Refresh Awareness specification moves all the work of
showing refresh progress and notifications into a specific snap.
To detect when that snap is available or not, snapd checks if
there is a snap with the snap-refresh-observe interface
connected. In that case, it presumes that it will manage all
the notifications, and won't send any by itself.

Unfortunately, only the "Pending refresh" notification checks
this case, while the "Refresh complete" notification doesn't.
This results in duplicated notifications.

This patch fixes this.