o/ifacestate: mock snapd.apparmor.service check
Interface manager asserts whether the snapd.apparmor.service is enabled, but the
tests were missing proper mocking of this helper. In this case, on hosts where
the snapd.apparmor.service was present but not installed, the check would fail
and thus ifacemgr would push a warning indicating a possible issue, which the
braks the tests, such as interfaceManagerSuite.TestBadInterfacesWarning.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>