tests: fix/optimise download timeout spread test
The download-timeout test was taking over 10min to complete due to a couple of
issues. The first one is that there was a typo in the environment variables
that controlled the measurement window so the timeout was still the default of
5min (the test downloads twice). Setting it correctly cut the execution time
for one of the downloads but not the other strangely. It turns out that
the test was creating a override file that was sorted before the other
existing override (systemd applies env vars in order).  Fixing this and
lowering the timeout to 15s cuts the running time from 10min to 30s. I wanted
to add a log when the env var was set so we could check it in the test but
this happens in an `init()` and I don't think the logger is set up by then.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>