tests: spread metadata used to filter and categorize tests
This change introduces metadate on spread tests, this metadate is
devided in features, groups and tags.

In all cases the matadata can be used to filter tests to be executed in
different circunstances, but also to know test coverage.

The file spread.meta.yaml includes the values allowed to include in the
tests. These values will be cross-checked in the ci to check the tests
integrity.