o/snapstate: ensure default provider waits for base in split refresh
In a split refresh, an app may refresh in parallel with the model base
(if it doesn't depend on it). However, it might pull in a default
provider that depends on the model. This change ensures that the default
provider waits for the model base to refresh and becomes available but
not for the entire essential snap task chain to finish (since that
requires a reboot).

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>