interfaces: add desktop-file-prefix plug attr to desktop interface
This plug attribute is needed to allow snaps to have custom desktop
file names.

Previously, snapd installed desktop files as “$snap_$app.desktop”.
This causes problems in a number of cases where desktop files are
expected to follow a particular format, or have a specific name.

The new plug attribute allows snaps to have custom desktop file
names (prefixed with the "desktop-file-prefix" attribute value).

Naming cannot be left uncontrolled, so the $MISSING constraint
is added to only allow overriding "desktop-file-prefix" through
the store. The process of requesting the snap declaration on the
store will help ensure that the prefixes are sensible.

More details can be found in the spec: SD170

Signed-off-by: Zeyad Gouda <zeyad.gouda@canonical.com>