tests: moving out google machines from us-east1-b
The region us-east1 is shared among project. In order to avoid reaching
the instances quota we are moving out to europe regions.
The idea of moving to europe is because the CI is in the UK.

From now, we will have 5 regions for snapd (4 in europe and 1 in the USA
for sru and nested)