osutil: switch to -u UID:GID for strace-static (#13950)
* osutil: switch to -u UID:GID for strace-static

This moves us off the custom patch and onto an upstream feature
we've heled develop. The feature is not released yet but the
patch has been integrated into the strace-static snap.

Jira: SNAPDENG-19870

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* tests: make strace test less fragile

The test tried to carefully match the error message to the version of strace
used, which in turn depends on the host OS. It's much easier to just check both
error mesages.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* tests: avoid systemctl kill which has issues on systemd 255

On systemd 255 systemctl kill fails after attempting to kill snapd.service with
the following message:

  $ sudo systemctl kill --signal=SIGKILL snapd.service
  Failed to kill unit snapd.service: Failed to send signal SIGKILL to auxiliary processes: Invalid argument

Kill the pid by hand to avoid triggering this behavior.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

---------

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>