packaging/snapd.mk: generate GNU build ID for Go binaries
The Go toolchain only generates a Go build ID, but GNU build IDs are expected to
present almost universally. Make sure to generate GNU build IDs for Go binaries
like we do in Fedora packaging, so that rpmbuild is happy. Note, that not all Go
toolchains we build with support -B gobuildid, hence populate the GNU build ID
with random bytes, but leave a note about a fix we need here.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>