cmd/snap-failure: do not attempt to revert snapd when reexec isn't supported by the systemd
The snap-failure helper is used as a general failure handler for failures of the
snapd.service unit. Its purpose is to trigger a revert of snapd snap to the
previous revision whenever a failure occurs. The logic asserting whether a snapd
refresh was in progress is privy to the actual snapd binary, however any attempt
to execute snapd binary from the snapd snap makes no sense if the system does
not support running snapd directly from the snap in the first place.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>