systemd: strip pre-release suffixes when comparing versions
Some tests were failing on debian-sid because we weren't coping with
systemd versions that contained a release candidate suffix. This deals
with it simply by ignoring which should be enough given that we only
compare them roughly to check if systemd is not too old.
```
error: cannot get logs: cannot get systemd version: cannot convert systemd
       version to number: 256~rc3
```

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>