o/snapstate: in hybrid systems re-refresh before reboot if possible (#13955)
* o/snapstate: rerefresh before reboot if possible

In hybrid systems where we do refreshes of apps before the reboot and
essential snaps after, check if a re-refresh is necessary after the apps
are done instead of waiting for all (and therefore a reboot). This is
possible because currently essential snaps don't use epochs and don't
require re-refreshes. If an app has a dependency that requires a reboot
(like the model base), then the check-rerefresh task needs to wait for
that refresh to complete. Since, if we're waiting for the reboot, there
isn't much to gain in resolving the app's dependencies to only wait
for those, the check-rerefresh preserves the old behavior in that case
to keep things simple.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

* o/snapstate: refactor and add unit test

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

---------

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>