snapdtool: proceed with trying reexec when SNAP_REEXEC=1 is set on an unsupported distro
Snapd normally stops trying to reexec when the current host distribution is not
in the list of supported ones. However, allow reexec to be attempt if
SNAP_REEXEC is explicitly set to a truthy value in the environment. This enables
testing of the reexec functionality on a wider range of distributions.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>