tests: simplify the pr-title check
The check-pr-title.py was created when we used travis (a long time ago).

Now, in github actions we can access to the pr title from ${{
github.event.pull_request.title }}

The idea of this change is to avoid checking in the github api and
simplify the check-pr-title.py tool.

This also is required to be able to use this check in private projects
where a unauthenticated api call is not allowed.

Also it is removed from the actions workflow the pr number variables
which are not used.