tests: test refresh mixing essential and apps in hybrid systems  (#13932)
* tests: test refresh mixing essential and apps in hybrid systems

Check that on a hybrid system, refreshes of both essential snaps and
apps and non-model bases don't make the apps wait for a reboot. In
particular check that the refresh of the kernel, gadget and model base
wait for a reboot. Check that apps that depend on the model base also
wait for it to be refreshed and that apps that don't can completely
refresh before the reboot.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

* tests: use correct model for system in muinstaller-real

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

---------

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>