i/prompting/requestprompts: record notice when request merged into prompt
When a new request is identical to an existing one and is merged into an
existing prompt, record a notice again for that prompt, despite that the
prompt contents itself have not changed.

The rationale for this is that if a client attempts to reply to a prompt
but sends a malformed response, snapd will respond with an error and the
prompt will not be resolved. If the client does not re-try to reply, the
application will be left hanging, waiting for a reply which will not
come.

If a user attempts to re-do the operation which triggered the prompt,
that request may be merged into an existing prompt, and the client will
not re-prompt the user or send a reply, leaving the application blocked
again, this time with no hope of a response, since the client has
received no notice for the prompt since the initial one, to which it
already sent a (malformed) reply.

Thus, by re-recording a notice when a request is merged into an existing
prompt, the user has a way of poking the client into taking action
again, and applications always cause a notice to be recorded when they
perform an action which should be handled by a prompt.

Of course, it is important that clients correctly handle error responses
from snapd and retry replying when appropriate. Re-recording a notice
when a request is merged into an existing prompt does not help in a
situation where an application is blocked on an operation which the
client thinks it has handled, and the application does not retry (which
is likely, given that the application itself is often blocked and
cannot/will not perform other actions until that syscall completes).

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>