run-checks: skip high-level tests requirement for interfaces with a standalone test
interfaces-many-{core,snap}-provided tests are doing minimal testing for
all interfaces. This became one of the slowest tests that it started to timeout.

A check is added to relax the requirement for high-level tests and
skip if a standalone test already exists for the interface which at
least connects and disconnects the interface.

The requirement can be relaxed further in the future to skip the
connect/disconnect checks allowing more tests to be skipped.

Signed-off-by: Zeyad Gouda <zeyad.gouda@canonical.com>