o/snapstate: refactor/optimise dependency setup in split refresh
Since the prerequisites code can deal with apps that depend on the model
base, we can remove the code that sets up explicit dependencies between
them. This also enables the app refresh to finish faster because the
prereq code will only wait for the base to link and not to fully refresh
(which requires the gadget/kernel to link, reboot, etc). Also, since
this means all non-essential snaps are guaranteed to finish before the
reboot, the rerefresh check can also always run before the reboot and
we can simplify some of the associated logic.
Also lowers the prerequisites timeout since this makes split refreshes
rely on it and 30 seconds is high.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>