i/prompting/requestprompts: include reason in notice data when prompt resolved
A prompt may only be resolved once, and the notice recorded when that
prompt is resolved is the final notice which will be associated with
that prompt ID. Therefore, it is safe to include notice data without
risk of signal loss.

Currently, prompt clients need to query the prompting API in order to
check whether an outstanding prompt was modified (due to some but not
all of its permissions being satisfied by a new rule) or fully resolved,
either due to a direct reply, a new rule satisfying all remaining
permissions, or the prompt DB being closed.

Since the notice recorded when a prompt is resolved is the final notice
for that prompt, we can instead include the "resolved" key in the
notice data to indicate to the client that the prompt has been resolved
without the need to query the prompting API. And the value for the
"resolved" key can be "replied", "satisfied", or "cancelled", giving
additional information to the prompt client about why the prompt was
resolved.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>