sandbox/apparmor: refactor parser feature test
The test was particularly expensive as it was testing all the combination of
known parser features, 2**N, with an entirely mocked parser written in shell.

Replace the test with a per-feature probe detector, a checker for all the
recognized features and separate smaller tests for not having any parser and
for using the internal parser.

Jira: https://warthogs.atlassian.net/browse/SNAPDENG-21937

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>