o/snapstate: speedup auto-refresh variants of split refresh tests
The auto-refresh variants of the split refresh tests were sometimes
failing when running with the race detector because besides refreshing
many snaps they also depend on tasks being retried (the prerequisites
and the check-rerefresh which unlike the other tests can't be disabled).
This changes those tests to refresh less snaps (which test scenarios
already covered in other tests) and lowers the timeouts.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>