cmd/libsnap-confine-private: add snap mount dir detection
Add a pair of functions to probe, memoize and return the value of snap mount
directory. The premise is to look at /snap in the initial mount namespace and
depending on its absence or presence and type, and if it is a symbolic link,
link target, either pick /snap, /var/lib/snapd/snap or die.

Jira: SNAPDENG-22335
Spec: SD-179

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>