tests: remove tiobe tics checks for prs
This is required because tics tools requires 10GB of disk space in the
runners to run. Tics creates a cache with tools used for processing the
changes that can go up to 10GB, and that makes our self-hosted runners
to go out of space.

In the future, when we create new runners, I'll condifer that disk space
constrain