o/snapstate: do not trigger fallback autorefresh schedule when managed with pending connect tasks
Auto-refresh schedule can only be set to "managed" when there is a snap with
"snapd-control" plug connected. When such snap is being refreshed, after
"setup-profiles" the connection would be broke and reestablished during
"auto-connect" when relevant "connect" tasks are injected. If snapd was to
restart before the connection is reestablished, it could initiate the
auto-refresh as the "managed" refresh schedule would appear to have been wrongly
established as there would be no snap plugging the "snapd-control" interface in
the state at this time.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>