o/hookstate/ctlcmd: fix queuing commands for default-configure hook
Multiple snaps could be installed in a single transaction
where all snap tasksets are in a single lane.

The old simplistic approach of looking up the first
"start-snap-services" fails when we have multiple
tasks for multiple snaps in the same lane.

A test is added to trigger this corner case, and the fix
just checks the snap name associated with the "start-snap-services"
tasks.

Signed-off-by: Zeyad Gouda <zeyad.gouda@canonical.com>