o/snapstate: make a managed refresh schedule not require any additional checks (#14107)
* o/snapstate: make a managed refresh schedule not require any additional checks

Drop the additional check to CanManageRefreshes() when the refresh schedule is
already set to 'managed'. This was originally a way to ensure that there is at
least one snap entitled to directly manage the refreshes or fall back to the
default auto-refresh schedule. However, the conditions in which the fallback
would be applied are incorrect and could lead to a situation when snapd would
trigger an auto-refresh even while a snap which is entitled to using a managed
refresh schedule is being refreshed (due to the snapd-control being temporarily
disconnected). On top of this, since the device was once switched to managed, it
clearly means that it was entitled to do so and it was intentional, hence we
should not accidentally break the expectations.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* o/devicestate: tweak comment around CanManageRefreshes()

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* o/configstate/configcore: tweak comments around managed refresh schedule

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* o/snapstate: refresh hints report managed refresh when schedule is set to managed

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* o/snapstate, o/devicestate: drop snapstate.CanManageRefreshes

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* overlord: add a durability test of connections during a refresh

* overlord: improve test for both old and fixed scenarios

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* overlord: tweak test comments

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* overlord: fix data race in unit tests

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

---------

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>
Co-authored-by: Philip Meulengracht <the_meulengracht@hotmail.com>